// src/lib/types/api.ts
export type PublicationType = "conference" | "article" | "chapter"

export interface Publication {
  id: number
  title: string
  year: number
  type: PublicationType
  pdfUrl?: string
  teamId?: number
  researcherId?: number
  createdAt?: string
}